[WIP] Splitting useTableControls into state and props hooks
